requirements.txt
# Kafka client — producer and consumer
kafka-python==2.0.2
 
# AWS S3 / MinIO client — same library works for both
boto3==1.34.34
 
# PostgreSQL driver
psycopg2-binary==2.9.9
 
# Data processing
pyspark==3.5.0
pandas==2.2.0
 
# Airflow — core + postgres connection
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.10.0
 
# Utilities
python-dotenv==1.0.1    # load config from .env files
retry==0.9.2            # retry decorator for flaky connections
 
# Testing and linting (used by CI/CD)
pytest==7.4.4
flake8==7.0.0